Add sync offline utils
